Adds bulk changes made alongside KidGame development, Bumps version to 1.0.0

+++ Runtime/Ticking/IUpdateable.cs
@@ -2,6 +2,6 @@
 {
     public interface IUpdateable
     {
-        void Update(double elapsedSeconds);
+        void Update(float elapsedSeconds);
     }
 }
